A simple calculator that can be used to perform simple calculations.
There is a simple python script that can be used to do simple calculations like Add, subtract, multiply and division.

The algorithm is simple and can be used to create calculation for beginners. hope you like the code and stay tuned for more updates.

And alo the code is not in turtle.